+++ cli/rt/01_web_util.js
@@ -41,19 +41,6 @@
     return Object.prototype.hasOwnProperty.call(obj, v);
   }
 
-  /** Returns whether o is iterable. */
-  function isIterable(
-    o,
-  ) {
-    // checks for null and undefined
-    if (o == null) {
-      return false;
-    }
-    return (
-      typeof (o)[Symbol.iterator] === "function"
-    );
-  }
-
   const objectCloneMemo = new WeakMap();
 
   function cloneArrayBuffer(
@@ -192,7 +179,6 @@
     requiredArguments,
     immutableDefine,
     hasOwnProperty,
-    isIterable,
     cloneValue,
     defineEnumerableProps,
     getHeaderValueParams,
