+++ op_crates/web/Cargo.toml
@@ -15,6 +15,8 @@ path = "lib.rs"
 
 [dependencies]
 deno_core = { version = "0.57.0", path = "../../core" }
+idna = "0.2.0"
+serde = { version = "1.0.116", features = ["derive"] }
 
 [dev-dependencies]
 futures = "0.3.5"
